BAEL-3290: Added test cases to exercise DI configuration
