--- spring-di/src/main/java/com/baeldung/dependency/ioc/Item.java
@@ -1,5 +1,0 @@
-package com.baeldung.dependency.ioc;
-
-public interface Item {
-
-}
