+++ spring-di/src/main/java/org/baeldung/store/AppConfig.java
@@ -1,4 +1,4 @@
-package com.baeldung.dependency.ioc;
+package org.baeldung.store;
 
 import org.springframework.context.annotation.Bean;
