+++ spring-di/src/main/java/org/baeldung/store/Item.java
@@ -1,0 +1,5 @@
+package org.baeldung.store;
+
+public interface Item {
+
+}
