+++ spring-di/src/main/java/org/baeldung/store/ItemImpl1.java
@@ -1,4 +1,4 @@
-package com.baeldung.dependency.ioc;
+package org.baeldung.store;
 
 public class ItemImpl1 implements Item {
