+++ spring-di/src/main/resources/ioc-context.xml
@@ -8,7 +8,7 @@
 
 	<bean id="item1" class="org.baeldung.store.ItemImpl1" />
 	<bean id="xml-store-by-constructor" class="org.baeldung.store.Store">
-		<constructor-arg type="ItemImpl1" index="0"	name="item" ref="item1" />
+		<constructor-arg type="Item" index="0"	name="item" ref="item1" />
 	</bean>
 	
 	<!-- Setter injection -->
